requirements.txt
# Core Dash Dependencies
dash>=2.14.1
dash-bootstrap-components==1.6.0
plotly>=5.15.0
# Data Processing
pandas>=2.1.1
numpy>=1.25.2
openpyxl>=3.1.2

# Database
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0

# File Processing
python-multipart>=0.0.6

# Development
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type Checking
mypy>=1.5.0
types-python-dateutil

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Production
gunicorn>=21.2.0
redis>=4.6.0

# Monitoring
sentry-sdk[flask]>=1.32.0

# Geographic/Map (optional)
geopandas>=0.13.0
shapely>=2.0.0
dash-leaflet>=1.0.0

# CSS Utilities (optional)
cssutils>=2.6.0

# YAML Configuration Support
PyYAML>=6.0
authlib>=1.2.1
flask-login>=0.6.2
python-jose>=3.3.0
cryptography>=41.0.3
flask-wtf>=1.2.1
Flask-Babel>=4.0.0
babel>=2.14.0